Create chatGPT_cli
This is a simple python program to integrate the OpenAI ChatGPT API into the cli of Kali Linux (I have not tested with any other kernels yet, sorry). This is the early stages of this development. The program has to main features, either use it as a regular chat or add a file as an argument and you can chat about the contents of the file. It's always a good idea to refer to the OpenAI API documentation or reach out to their support for the most up-to-date information on file type and size limitations. This program was created with ChatGPT.